+++ features/admin.core.v1/store/reducers/config.ts
@@ -184,6 +184,7 @@ export const commonConfigReducerInitialState: CommonConfigReducerStateInterface<
             smsTemplates: "",
             tenantAssociationApi: "",
             tenantManagementApi: "",
+            tenantMoesifDashboardInfo: "",
             tenantSubscriptionApi: "",
             tenants: "",
             updateSecret: "",