+++ features/admin.extensions.v1/public-api.ts
@@ -16,8 +16,8 @@
  * under the License.
  */
 
+export * from "./configs";
 export * from "./extensions-manager";
 export * from "./models";
-export * from "./configs";
 
 export { featureGateConfig } from "./configs/feature-gate";